:ambulance: delete comment fixed

+++ server/data/comments.js
@@ -130,14 +130,24 @@ const updateComment = async (id, comment, userId, name) => {
   }
 };
 
-const deleteComment = async (id) => {
+const deleteComment = async (id, storyId) => {
   let comment;
   if (!uuidValidate(id)) {
     throw TypeError('Id is of invalid type');
   }
+  if (!uuidValidate(storyId)) {
+    throw TypeError('Story Id is of invalid type');
+  }
   try {
     const commentsCollection = await commentSchema();
     comment = await commentsCollection.deleteOne({ _id: id });
+    const storiesCollection = await storiesSchema();
+    await storiesCollection.updateOne(
+      {
+        _id: storyId,
+      },
+      { $pull: { comments: id } }
+    );
     return comment;
   } catch (error) {
     throw Error(error.message);
